requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
python-decouple>=3.8
gunicorn>=21.0
psycopg2-binary>=2.9
pytest>=7.4
pytest-django>=4.7